mcp_client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.69.0",
    "mcp>=1.16.0",
    "python-dotenv>=1.1.1",
]
